api/Python/requirements.txt
azure-common==1.1.28
azure-core==1.30.0
azure-cognitiveservices-speech==1.33.0
azure-cosmos==4.6.0
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.1
backoff==2.2.1
bs4==0.0.1
cssutils==2.9.0
fake-useragent==1.2.1
Faker==19.1.0
flask==3.0.2
langchain==0.1.13
langchain-core==0.1.34
langchain-experimental==0.0.55
langchain-community==0.0.29
langsmith==0.1.33
llama-index==0.9.40
llama_hub==0.0.50
lxml==5.1.0
msal==1.24.0
msal-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
openai==1.12.0
openpyxl==3.1.2
pandas==1.5.3
pathos==0.2.9
pdfminer.six==20221105
Pillow==10.1.0
pytz==2022.7.1
pypdf==4.0.1
ratelimit==2.2.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
tenacity==8.2.2
tiktoken==0.6.0
tqdm==4.66.1
urllib3==2.0.4
pyodbc==5.0.1
azure-functions-durable==1.2.8
azure-functions==1.18.0
yfinance==0.2.36
azure-ai-formrecognizer==3.3.2
boto3==1.34.34
unstructured[all-docs]==0.12.5
pymupdf==1.23.26
bcolors==1.0.4
azure-ai-documentintelligence==1.0.0b2
langchain-openai==0.1.1
langchainhub==0.1.15
langchain-pinecone==0.0.3